Add files via upload
This file contains the metrics we worked with on Thursday because to the best of my knowledge they are not in any of the other files. The charts are created in this file. Bokeh doesn't create the graph directly inside the notebook, but rather it sets up a webpage that contains the graph.